[JENKINS-62526] Split JenkinsLogs
