+++ src/main/java/com/cloudbees/jenkins/support/impl/CustomLogs.java
@@ -1,0 +1,217 @@
+package com.cloudbees.jenkins.support.impl;
+
+import com.cloudbees.jenkins.support.api.Component;
+import com.cloudbees.jenkins.support.api.Container;
+import com.cloudbees.jenkins.support.api.FileContent;
+import com.cloudbees.jenkins.support.api.SupportContext;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import hudson.Extension;
+import hudson.init.Terminator;
+import hudson.logging.LogRecorder;
+import hudson.model.PeriodicWork;
+import hudson.security.Permission;
+import hudson.util.CopyOnWriteList;
+import hudson.util.io.RewindableFileOutputStream;
+import hudson.util.io.RewindableRotatingFileOutputStream;
+import io.jenkins.lib.support_log_formatter.SupportLogFormatter;
+import jenkins.model.Jenkins;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+import java.util.logging.Handler;
+import java.util.logging.Level;
+import java.util.logging.LogRecord;
+import java.util.logging.Logger;
+import java.util.logging.StreamHandler;
+
+/**
+ * Custom Log recorders files from the master.
+ */
+@Extension(ordinal = 100.0)
+public class CustomLogs extends Component {
+
+    private static final int MAX_ROTATE_LOGS = Integer.getInteger(CustomLogs.class.getName() + ".MAX_ROTATE_LOGS", 9);
+    private final Map<String,LogRecorder> logRecorders = Jenkins.get().getLog().logRecorders;
+    private final File customLogs = new File(getLogsRoot(), "custom");
+
+    @NonNull
+    @Override
+    public Set<Permission> getRequiredPermissions() {
+        return Collections.singleton(Jenkins.ADMINISTER);
+    }
+
+    @NonNull
+    @Override
+    public String getDisplayName() {
+        return "Master Custom Log Recorders";
+    }
+
+    @Override
+    public void start(@NonNull SupportContext context) {
+        Logger.getLogger("").addHandler(new CustomHandler());
+    }
+
+    @Override
+    public void addContents(@NonNull Container result) {
+        addLogRecorders(result);
+    }
+
+    /**
+     * Dumps the content of {@link LogRecorder}, which is the groups of loggers configured
+     * by the user. The contents are also ring buffer and only remembers recent 256 or so entries.
+     */
+    private void addLogRecorders(Container result) {
+        for (Map.Entry<String, LogRecorder> entry : logRecorders.entrySet()) {
+            String name = entry.getKey();
+            String entryName = "nodes/master/logs/custom/{0}.log"; // name to be filtered in the bundle
+            File storedFile = new File(customLogs, name + ".log");
+            if (storedFile.isFile()) {
+                result.add(new FileContent(entryName, new String[]{name}, storedFile));
+            } else {
+                // Was not stored for some reason; fine, just load the memory buffer.
+                final LogRecorder recorder = entry.getValue();
+                result.add(new LogRecordContent(entryName, new String[]{name}) {
+                    @Override
+                    public Iterable<LogRecord> getLogRecords() {
+                        return recorder.getLogRecords();
+                    }
+                });
+            }
+        }
+    }
+
+    /**
+     * Returns the root directory for logs (historically always found under <code>$JENKINS_HOME/logs</code>.
+     * Configurable since Jenkins 2.114.
+     *
+     * @see hudson.triggers.SafeTimerTask#LOGS_ROOT_PATH_PROPERTY
+     * @return the root directory for logs.
+     */
+    private File getLogsRoot() {
+        final String overriddenLogsRoot = System.getProperty("hudson.triggers.SafeTimerTask.logsTargetDir");
+        if (overriddenLogsRoot == null) {
+            return new File(Jenkins.get().getRootDir(), "logs");
+        } else {
+            return new File(overriddenLogsRoot);
+        }
+    }
+
+    @SuppressFBWarnings(value="SIC_INNER_SHOULD_BE_STATIC_NEEDS_THIS", justification="customLogs is not static, so this is a bug in FB")
+    private final class LogFile {
+        private final RewindableRotatingFileOutputStream stream;
+        private final Handler handler;
+        private int count;
+        @SuppressFBWarnings(value="RV_RETURN_VALUE_IGNORED_BAD_PRACTICE", justification="if mkdirs fails, will just get a stack trace later")
+        LogFile(String name) throws IOException {
+            customLogs.mkdirs();
+            stream = new RewindableRotatingFileOutputStream(new File(customLogs, name + ".log"), MAX_ROTATE_LOGS);
+            // TODO there is no way to avoid rotating when first opened; if .rewind is skipped, the file is just truncated
+            stream.rewind();
+            handler = new StreamHandler(stream, new SupportLogFormatter());
+            handler.setLevel(Level.ALL);
+            count = 0;
+        }
+        void publish(LogRecord record) throws IOException {
+            boolean rewind = false;
+            synchronized (this) {
+                if (count++ > 9999) { // make sure it does not get enormous during a single session
+                    count = 0;
+                    rewind = true;
+                }
+            }
+            if (rewind) {
+                stream.rewind();
+            }
+            handler.publish(record);
+            LogFlusher.scheduleFlush(handler);
+        }
+    }
+
+    private final class CustomHandler extends Handler {
+
+        private final Map<String,LogFile> logFiles = new HashMap<String,LogFile>();
+
+        /** JENKINS-27669: try to preload classes that will be needed by {@link #publish} */
+        CustomHandler() {
+            Arrays.hashCode(new Class<?>[] {
+                    Map.Entry.class,
+                    LogRecorder.class,
+                    LogRecorder.Target.class,
+                    LogFile.class,
+                    RewindableFileOutputStream.class,
+                    RewindableRotatingFileOutputStream.class,
+                    StreamHandler.class,
+                    SupportLogFormatter.class,
+                    LogFlusher.class,
+                    CopyOnWriteList.class,
+                    PrintWriter.class,
+                    Throwable.class,
+            });
+        }
+
+        @Override public void publish(LogRecord record) {
+            for (Map.Entry<String,LogRecorder> entry : logRecorders.entrySet()) {
+                for (LogRecorder.Target target : entry.getValue().targets) {
+                    if (target.includes(record)) {
+                        try {
+                            String name = entry.getKey();
+                            LogFile logFile;
+                            synchronized (logFiles) {
+                                logFile = logFiles.get(name);
+                                if (logFile == null) {
+                                    logFile = new LogFile(name);
+                                    logFiles.put(name, logFile);
+                                }
+                            }
+                            logFile.publish(record);
+                        } catch (IOException x) {
+                            x.printStackTrace(); // TODO probably unsafe to log this
+                        }
+                    }
+                }
+            }
+        }
+
+        @Override public void flush() {}
+
+        @Override public void close() throws SecurityException {}
+
+    }
+
+    @Extension public static final class LogFlusher extends PeriodicWork {
+
+        private static final Set<Handler> unflushedHandlers = new HashSet<Handler>();
+
+        static synchronized void scheduleFlush(Handler h) {
+            unflushedHandlers.add(h);
+        }
+
+        @Override public long getRecurrencePeriod() {
+            return 3000; // 3s
+        }
+
+        @Override protected void doRun() throws Exception {
+            flush();
+        }
+
+        @Terminator public static void flush() {
+            Handler[] handlers;
+            synchronized (LogFlusher.class) {
+                handlers = unflushedHandlers.toArray(new Handler[unflushedHandlers.size()]);
+                unflushedHandlers.clear();
+            }
+            for (Handler h : handlers) {
+                h.flush();
+            }
+        }
+
+    }
+}
